+++ random.js
@@ -1,0 +1,5 @@
+export const getRandom = (max) => {
+    return Math.ceil(Math.random() * max);
+}
+
+export default getRandom;
